+++ Rabbit-Island/Entities/ICreature.cs
@@ -33,7 +33,7 @@ namespace Rabbit_Island.Entities
 
         public Creature? PregnantWith { get; }
 
-        public void LoseHealth(float damage);
+        public void LoseHealth(double damage);
 
         public bool CanMate { get; }
     }
